djangoSocketServer/requirements.txt
Django
djangorestframework
eventlet
py-eureka-client
python-socketio
pyjwt